+++ include/cvode/cvode_diag.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.3 $
- * $Date: 2006-11-29 00:05:06 $
+ * $Revision: 1.4 $
+ * $Date: 2010-12-01 22:10:38 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Scott D. Cohen, Alan C. Hindmarsh and
  *                Radu Serban @ LLNL
@@ -70,7 +70,7 @@ SUNDIALS_EXPORT int CVDiag(void *cvode_mem);
 
 SUNDIALS_EXPORT int CVDiagGetWorkSpace(void *cvode_mem, long int *lenrwLS, long int *leniwLS);
 SUNDIALS_EXPORT int CVDiagGetNumRhsEvals(void *cvode_mem, long int *nfevalsLS);
-SUNDIALS_EXPORT int CVDiagGetLastFlag(void *cvode_mem, int *flag);
+SUNDIALS_EXPORT int CVDiagGetLastFlag(void *cvode_mem, long int *flag);
 
 /*
  * -----------------------------------------------------------------
@@ -79,7 +79,7 @@ SUNDIALS_EXPORT int CVDiagGetLastFlag(void *cvode_mem, int *flag);
  * -----------------------------------------------------------------
  */
   
-SUNDIALS_EXPORT char *CVDiagGetReturnFlagName(int flag);
+SUNDIALS_EXPORT char *CVDiagGetReturnFlagName(long int flag);
 
 /*
  * -----------------------------------------------------------------
