+++ include/cvode/cvode_spils.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.9 $
- * $Date: 2008-09-10 22:36:38 $
+ * $Revision: 1.10 $
+ * $Date: 2010-12-01 22:10:38 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Scott D. Cohen, Alan C. Hindmarsh and
  *                Radu Serban @ LLNL
@@ -347,7 +347,7 @@ SUNDIALS_EXPORT int CVSpilsGetNumLinIters(void *cvode_mem, long int *nliters);
 SUNDIALS_EXPORT int CVSpilsGetNumConvFails(void *cvode_mem, long int *nlcfails);
 SUNDIALS_EXPORT int CVSpilsGetNumJtimesEvals(void *cvode_mem, long int *njvevals);
 SUNDIALS_EXPORT int CVSpilsGetNumRhsEvals(void *cvode_mem, long int *nfevalsLS); 
-SUNDIALS_EXPORT int CVSpilsGetLastFlag(void *cvode_mem, int *flag);
+SUNDIALS_EXPORT int CVSpilsGetLastFlag(void *cvode_mem, long int *flag);
 
 /*
  * -----------------------------------------------------------------
@@ -356,7 +356,7 @@ SUNDIALS_EXPORT int CVSpilsGetLastFlag(void *cvode_mem, int *flag);
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT char *CVSpilsGetReturnFlagName(int flag);
+SUNDIALS_EXPORT char *CVSpilsGetReturnFlagName(long int flag);
 
 #ifdef __cplusplus
 }
