+++ include/ida/ida_bbdpre.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.7 $
- * $Date: 2007-11-26 16:19:59 $
+ * $Revision: 1.8 $
+ * $Date: 2010-12-01 22:14:08 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Alan C. Hindmarsh, Radu Serban and
  *                Aaron Collier @ LLNL
@@ -136,7 +136,7 @@ extern "C" {
  * -----------------------------------------------------------------
  */
 
-typedef int (*IDABBDLocalFn)(int Nlocal, realtype tt,
+typedef int (*IDABBDLocalFn)(long int Nlocal, realtype tt,
 			     N_Vector yy, N_Vector yp, N_Vector gval,
 			     void *user_data);
 
@@ -168,7 +168,7 @@ typedef int (*IDABBDLocalFn)(int Nlocal, realtype tt,
  * -----------------------------------------------------------------
  */
 
-typedef int (*IDABBDCommFn)(int Nlocal, realtype tt,
+typedef int (*IDABBDCommFn)(long int Nlocal, realtype tt,
 			    N_Vector yy, N_Vector yp,
 			    void *user_data);
 
@@ -215,9 +215,9 @@ typedef int (*IDABBDCommFn)(int Nlocal, realtype tt,
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT int IDABBDPrecInit(void *ida_mem, int Nlocal,
-                                   int mudq, int mldq,
-                                   int mukeep, int mlkeep,
+SUNDIALS_EXPORT int IDABBDPrecInit(void *ida_mem, long int Nlocal,
+                                   long int mudq, long int mldq,
+                                   long int mukeep, long int mlkeep,
                                    realtype dq_rel_yy,
                                    IDABBDLocalFn Gres, IDABBDCommFn Gcomm);
 
@@ -244,7 +244,7 @@ SUNDIALS_EXPORT int IDABBDPrecInit(void *ida_mem, int Nlocal,
  */
 
 SUNDIALS_EXPORT int IDABBDPrecReInit(void *ida_mem,
-				     int mudq, int mldq,
+				     long int mudq, long int mldq,
 				     realtype dq_rel_yy);
 
 /*
