+++ include/kinsol/kinsol_band.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.4 $
- * $Date: 2008-04-18 19:42:38 $
+ * $Revision: 1.5 $
+ * $Date: 2010-12-01 22:16:17 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -48,7 +48,7 @@ extern "C" {
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT int KINBand(void *kinmem, int N, int mupper, int mlower);
+SUNDIALS_EXPORT int KINBand(void *kinmem, long int N, long int mupper, long int mlower);
 
 #ifdef __cplusplus
 }
